Updated server/client connections with buffer size adjustments
Before sending over the large json objects (the initial data from the
server and the final moves from the client), they first send a message
with the minimum number of bytes needed in the buffer to receive the
message.

Both sides wait 2 seconds between sending the initial buffer size
message and the actual message. To account for this, I have the server
allows for a maximum time limit of 122 seconds for the player to decide
their move